Finished extracting all data for survey

+++ src/main/java/com/voting/survey_host/rowMapper/SurveyRowMapper.java
@@ -17,7 +17,7 @@ public class SurveyRowMapper implements RowMapper<Survey> {
     @Override
     public Survey mapRow(ResultSet rs, int rowNum) throws SQLException {
         Survey survey = new Survey();
-        survey.setSurveyId((int) rs.getLong("survey_id"));
+        survey.setSurveyId(rs.getLong("survey_id"));
         survey.setHostUsername(hostname);
         survey.setTitle(rs.getString("title"));
         return survey;
